chore: Start testing pyspark integration
It looks like we have a basic bug
