+++ .github/workflows/python_build.yml
@@ -132,3 +132,46 @@ jobs:
       run: |
         source venv/bin/activate
         make build-documentation
+
+  test-pyspark:
+    name: PySpark Integration Tests
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v2
+
+    - uses: actions/cache@v2
+      with:
+        path: |
+          ~/.cargo/bin/
+          ~/.cargo/registry/index/
+          ~/.cargo/registry/cache/
+          ~/.cargo/git/db/
+          target/
+        key: python-pyspark-test-${{ hashFiles('**/Cargo.lock', '**/Cargo.toml') }}
+
+    - name: Install latest nightly
+      uses: actions-rs/toolchain@v1
+      with:
+          toolchain: stable
+          override: true
+          components: rustfmt, clippy
+
+    - uses: actions/setup-python@v3
+      with:
+        python-version: '3.10' 
+    
+    - uses: actions/setup-java@v1
+      with:
+        java-version: '11'
+
+    - name: Build and install deltalake
+      run: |
+        pip install virtualenv
+        virtualenv venv
+        source venv/bin/activate
+        make develop-pyspark
+
+    - name: Run tests
+      run: |
+        source venv/bin/activate
+        make test-pyspark
